Update post-release workflow
Because tm-compsec has fixed the issues about SSH, we can now call
workflows in other projects through the Bitrise-native step
`trigger-bitrise-workflow`. This will call a workflow in
glia-ios-podspecs to create a new podspec with the new version. There
is no need to create a GitHub Action for this because the ios-sdk
release process automatically creates a tag in ios-bundle, which will
then trigger this post-release workflow.

MOB-1589