Add workflow for manually updating the widgets SDK dependencies
As we now depend on Cocoapods for the release, we unfortunately will
need to manually execute a workflow after Cocoapods is done with it.
This workflow will execute yet another workflow in the widgets, which
will run `pod uodate`. A GitHub Action to execute this through
GitHub will come in a future PR.

MOB-1791