Update Widgets SDK when Core SDK is released
After the Core SDK is released successfully in Cocoapods, the new
`post-release` workflow should be executed. This workflow goes into
the Package.swift file, gets the checksum and the version, and then
calls the `update-dependencies` workflow in the ios-sdk-widgets
project. This wokflow will execute some new stuff that will make sure
to update correctly dependencies and checksums.

MOB-1792